+++ tests/unit/utils/test_transforms.py
@@ -1,5 +1,4 @@
 import pytorch_testing_utils as ptu
-import torch
 
 import pystiche.image.transforms.functional as F
 from pystiche_papers import utils
@@ -11,8 +10,8 @@ def test_OptionalGrayscaleToFakegrayscale_no_grayscale(content_image):
     ptu.assert_allclose(actual, content_image)
 
 
-def test_OptionalGrayscaleToFakegrayscale():
-    image = torch.rand((1, 1, 32, 32))
+def test_OptionalGrayscaleToFakegrayscale(content_image):
+    image = F.rgb_to_grayscale(content_image)
     transform = utils.OptionalGrayscaleToFakegrayscale()
     actual = transform(image)
     desired = F.grayscale_to_fakegrayscale(image)
